Remove warning from safe-transform

+++ tasks/safe-transform/test.cpp
@@ -129,7 +129,7 @@ TEST_CASE("Copy fails") {
         data.begin(), data.end(), [](const Int&) { return true; },
         [](Int& val) { val.x *= val.x; });
     for (size_t i = 0; i < data.size(); ++i) {
-        REQUIRE(data[i].x == (i + 1) * (i + 1));
+        REQUIRE(static_cast<size_t>(data[i].x) == (i + 1) * (i + 1));
     }
 }
